Splitted DAGs into main entities elt
